Create unit test for registerBuiltInActionTypes and move common code
