+++ x-pack/legacy/plugins/actions/server/builtin_action_types/pagerduty.test.ts
@@ -9,25 +9,15 @@ jest.mock('./lib/post_pagerduty', () => ({
 }));
 
 import { ActionType, Services, ActionTypeExecutorOptions } from '../types';
-import { ActionsConfigurationUtilities } from '../actions_config';
-import { ActionTypeRegistry } from '../action_type_registry';
-import { taskManagerMock } from '../../../task_manager/task_manager.mock';
-import { encryptedSavedObjectsMock } from '../../../encrypted_saved_objects/server/plugin.mock';
 import { validateConfig, validateSecrets, validateParams } from '../lib';
 import { SavedObjectsClientMock } from '../../../../../../src/core/server/mocks';
 import { postPagerduty } from './lib/post_pagerduty';
-import { registerBuiltInActionTypes } from './index';
+import { createActionTypeRegistry } from './index.test';
 
 const postPagerdutyMock = postPagerduty as jest.Mock;
 
 const ACTION_TYPE_ID = '.pagerduty';
 const NO_OP_FN = () => {};
-const MOCK_KIBANA_CONFIG_UTILS: ActionsConfigurationUtilities = {
-  isWhitelistedHostname: _ => true,
-  isWhitelistedUri: _ => true,
-  ensureWhitelistedHostname: _ => {},
-  ensureWhitelistedUri: _ => {},
-};
 
 const services: Services = {
   log: NO_OP_FN,
@@ -35,36 +25,15 @@ const services: Services = {
   savedObjectsClient: SavedObjectsClientMock.create(),
 };
 
-function getServices(): Services {
-  return services;
-}
-
 let actionType: ActionType;
-let actionTypeRegistry: ActionTypeRegistry;
-
-const mockEncryptedSavedObjectsPlugin = encryptedSavedObjectsMock.create();
 
 beforeAll(() => {
-  actionTypeRegistry = new ActionTypeRegistry({
-    getServices,
-    isSecurityEnabled: true,
-    taskManager: taskManagerMock.create(),
-    encryptedSavedObjectsPlugin: mockEncryptedSavedObjectsPlugin,
-    spaceIdToNamespace: jest.fn().mockReturnValue(undefined),
-    getBasePath: jest.fn().mockReturnValue(undefined),
-  });
-  registerBuiltInActionTypes(actionTypeRegistry, MOCK_KIBANA_CONFIG_UTILS);
+  const actionTypeRegistry = createActionTypeRegistry();
   actionType = actionTypeRegistry.get(ACTION_TYPE_ID);
 });
 
 beforeEach(() => {
   services.log = NO_OP_FN;
-});
-
-describe('action registation', () => {
-  test('should be successful', () => {
-    expect(actionTypeRegistry.has(ACTION_TYPE_ID)).toEqual(true);
-  });
 });
 
 describe('get()', () => {
