+++ x-pack/legacy/plugins/actions/server/builtin_action_types/server_log.test.ts
@@ -5,23 +5,12 @@
  */
 
 import { ActionType, Services } from '../types';
-import { ActionsConfigurationUtilities } from '../actions_config';
-import { ActionTypeRegistry } from '../action_type_registry';
-import { taskManagerMock } from '../../../task_manager/task_manager.mock';
-import { encryptedSavedObjectsMock } from '../../../encrypted_saved_objects/server/plugin.mock';
 import { validateParams } from '../lib';
 import { SavedObjectsClientMock } from '../../../../../../src/core/server/mocks';
-
-import { registerBuiltInActionTypes } from './index';
+import { createActionTypeRegistry } from './index.test';
 
 const ACTION_TYPE_ID = '.server-log';
 const NO_OP_FN = () => {};
-const MOCK_KIBANA_CONFIG_UTILS: ActionsConfigurationUtilities = {
-  isWhitelistedHostname: _ => true,
-  isWhitelistedUri: _ => true,
-  ensureWhitelistedHostname: _ => {},
-  ensureWhitelistedUri: _ => {},
-};
 
 const services: Services = {
   log: NO_OP_FN,
@@ -29,52 +18,26 @@ const services: Services = {
   savedObjectsClient: SavedObjectsClientMock.create(),
 };
 
-function getServices(): Services {
-  return services;
-}
-
-let actionTypeRegistry: ActionTypeRegistry;
-
-const mockEncryptedSavedObjectsPlugin = encryptedSavedObjectsMock.create();
+let actionType: ActionType;
 
 beforeAll(() => {
-  actionTypeRegistry = new ActionTypeRegistry({
-    getServices,
-    isSecurityEnabled: true,
-    taskManager: taskManagerMock.create(),
-    encryptedSavedObjectsPlugin: mockEncryptedSavedObjectsPlugin,
-    spaceIdToNamespace: jest.fn().mockReturnValue(undefined),
-    getBasePath: jest.fn().mockReturnValue(undefined),
-  });
-  registerBuiltInActionTypes(actionTypeRegistry, MOCK_KIBANA_CONFIG_UTILS);
+  const actionTypeRegistry = createActionTypeRegistry();
+  actionType = actionTypeRegistry.get(ACTION_TYPE_ID);
+  expect(actionType).toBeTruthy();
 });
 
 beforeEach(() => {
   services.log = NO_OP_FN;
 });
 
-describe('action is registered', () => {
-  test('gets registered with builtin actions', () => {
-    expect(actionTypeRegistry.has(ACTION_TYPE_ID)).toEqual(true);
-  });
-});
-
 describe('get()', () => {
   test('returns action type', () => {
-    const actionType = actionTypeRegistry.get(ACTION_TYPE_ID);
     expect(actionType.id).toEqual(ACTION_TYPE_ID);
     expect(actionType.name).toEqual('server-log');
   });
 });
 
 describe('validateParams()', () => {
-  let actionType: ActionType;
-
-  beforeAll(() => {
-    actionType = actionTypeRegistry.get(ACTION_TYPE_ID);
-    expect(actionType).toBeTruthy();
-  });
-
   test('should validate and pass when params is valid', () => {
     expect(validateParams(actionType, { message: 'a message' })).toEqual({
       message: 'a message',
@@ -123,7 +86,6 @@ describe('execute()', () => {
     const mockLog = jest.fn().mockResolvedValueOnce({ success: true });
 
     services.log = mockLog;
-    const actionType = actionTypeRegistry.get(ACTION_TYPE_ID);
     const id = 'some-id';
     await actionType.executor({
       id,
