+++ x-pack/package.json
@@ -155,6 +155,7 @@
     "pdf-image": "2.0.0",
     "pdfjs-dist": "^2.0.943",
     "pixelmatch": "4.0.2",
+    "pre-commit": "^1.2.2",
     "proxyquire": "1.8.0",
     "react-docgen-typescript-loader": "^3.1.1",
     "react-hooks-testing-library": "^0.3.8",
